Move TileGraph to new graph/internal
Only used for testing, so don't export it to the public.
